misc/HtmlTestSuite/results/test-1310.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>Custom elements: Range APIs should invoke constructor in tree order</title>
<meta href="mailto:echen@mozilla.com" name="author" title="Edgar Chen"/>
<link href="https://html.spec.whatwg.org/multipage/scripting.html#concept-upgrade-an-element" rel="help"/>
<link href="https://dom.spec.whatwg.org/#concept-create-element" rel="help"/>
</head><body><line href="https://dom.spec.whatwg.org/#concept-range-extract" rel="help">
<line href="https://dom.spec.whatwg.org/#concept-range-clone" rel="help">
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>


<div id="log"/>

<c-e data-index="root">
  <c-e data-index="root-0">
    <c-e data-index="root-0-0">
      <c-e data-index="root-0-0-0"/>
      <span id="start"/>
    </c-e>
  </c-e>
  <c-e data-index="root-1"/>
  <span id="end"/>
</c-e>

<script>

var logs = [];
class CE extends HTMLElement {
  constructor() {
    super();
    logs.push(this.dataset.index);
  }
}
customElements.define('c-e', CE);

function getRange() {
  const range = new Range();
  range.setStart(document.getElementById('start'), 0);
  range.setEnd(document.getElementById('end'), 0);
  return range;
}

test(function () {
  // Clear log for testing.
  logs = [];
  getRange().cloneContents();
  assert_array_equals(logs, ['root-0', 'root-0-0', 'root-1']);
}, 'Range.cloneContents should invoke constructor in tree order');

test(function () {
  // Clear log for testing.
  logs = [];
  getRange().extractContents();
  assert_array_equals(logs, ['root-0', 'root-0-0']);
}, 'Range.extractContents should invoke constructor in tree order');

</script>
</line></line></body>
</html>
